+++ .gitlab/notify.yml
@@ -7,6 +7,7 @@ include:
 
 notify-on-tagged-success:
   extends: .slack-notifier-base
+  stage: notify
   rules:
     !reference [.on_deploy_stable_or_beta_repo_branch]
   dependencies: []
@@ -17,23 +18,45 @@ notify-on-tagged-success:
 
 notify:
   extends: .slack-notifier-base
+  stage: notify
   rules:
-    !reference [.on_a7]
+    !reference [.on_main_or_release_branch_or_deploy_always]
   dependencies: []
   script:
     - set +x
     - export GITLAB_TOKEN=$(aws ssm get-parameter --region us-east-1 --name ci.datadog-agent.gitlab_read_api_token --with-decryption --query "Parameter.Value" --out text)
-    - export DD_API_KEY=$(aws ssm get-parameter --region us-east-1 --name ci.datadog-agent.datadog_api_key_org2 --with-decryption --query "Parameter.Value" --out text)
-    - python3 -m pip install -r tasks/libs/requirements-notifications.txt -r tasks/libs/requirements-datadog-api.txt
+    - python3 -m pip install -r tasks/libs/requirements-notifications.txt
     - |
-      # Do not send notifications and statistics if this is a child pipeline of another repo
+      # Do not send notifications if this is a child pipeline of another repo
       # The triggering repo should already have its own notification system
       if [ "$CI_PIPELINE_SOURCE" != "pipeline" ]; then
         if [ "$DEPLOY_AGENT" = "true" ]; then
-          invoke -e pipeline.notify --notification-type "deploy" --send-stats
+          invoke -e pipeline.notify --notification-type "deploy"
         else
-          invoke -e pipeline.notify --notification-type "merge" --send-stats
+          invoke -e pipeline.notify --notification-type "merge"
         fi
       else
         echo "This pipeline was triggered by another repository, skipping notification."
+      fi
+
+send_pipeline_stats:
+  stage: notify
+  # Using the buildimages image and not slack-notifier because we need Python 3.7+
+  image: 486234852809.dkr.ecr.us-east-1.amazonaws.com/ci/datadog-agent-buildimages/deb_x64:$DATADOG_AGENT_BUILDIMAGES
+  tags: ["arch:amd64"]
+  rules:
+    !reference [.on_a7]
+  dependencies: []
+  script:
+    - source /root/.bashrc && conda activate ddpy3
+    - set +x
+    - export GITLAB_TOKEN=$(aws ssm get-parameter --region us-east-1 --name ci.datadog-agent.gitlab_read_api_token --with-decryption --query "Parameter.Value" --out text)
+    - export DD_API_KEY=$(aws ssm get-parameter --region us-east-1 --name ci.datadog-agent.datadog_api_key_org2 --with-decryption --query "Parameter.Value" --out text)
+    - python3 -m pip install -r tasks/libs/requirements-stats.txt
+    - |
+      # Do not send statistics if this is a child pipeline of another repo
+      if [ "$CI_PIPELINE_SOURCE" != "pipeline" ]; then
+          invoke -e pipeline.send-stats
+      else
+        echo "This pipeline was triggered by another repository, skipping statistics."
       fi
