+++ tasks/libs/requirements-stats.txt
@@ -1,1 +1,2 @@
+invoke==1.7.1
 datadog-api-client==2.3.0
