+++ tasks/pipeline.py
@@ -13,13 +13,9 @@ from invoke.exceptions import Exit
 from .libs.common.color import color_message
 from .libs.common.gitlab import Gitlab, get_gitlab_bot_token, get_gitlab_token
 from .libs.datadog_api import create_count, send_metrics
-from .libs.pipeline_notifications import (
-    base_message,
-    find_job_owners,
-    get_failed_jobs,
-    get_failed_tests,
-    send_slack_message,
-)
+from .libs.pipeline_data import get_failed_jobs
+from .libs.pipeline_notifications import base_message, find_job_owners, get_failed_tests, send_slack_message
+from .libs.pipeline_stats import get_failed_jobs_stats
 from .libs.pipeline_tools import (
     FilteredOutException,
     cancel_pipelines_with_confirmation,
@@ -348,17 +344,9 @@ Please check for typos in the JOBOWNERS file and/or add them to the Github <-> S
 """
 
 
-def generate_failure_data():
-    project_name = "DataDog/datadog-agent"
-    failed_jobs, job_failure_stats = get_failed_jobs(project_name, os.getenv("CI_PIPELINE_ID"))
-
-    messages_to_send = generate_failure_messages(project_name, failed_jobs)
-
-    return messages_to_send, job_failure_stats
-
-
 def generate_failure_messages(project_name, failed_jobs):
     all_teams = "@DataDog/agent-all"
+
     # Generate messages for each team
     messages_to_send = defaultdict(lambda: TeamMessage())
     messages_to_send[all_teams] = SlackMessage(jobs=failed_jobs)
@@ -446,17 +434,17 @@ def trigger_child_pipeline(_, git_ref, project_name, variables="", follow=True):
 
 
 @task
-def notify(_, notification_type="merge", print_to_stdout=False, send_stats=False):
+def notify(_, notification_type="merge", print_to_stdout=False):
     """
     Send notifications for the current pipeline. CI-only task.
     Use the --print-to-stdout option to test this locally, without sending
     real slack messages.
-    Use the --send-stats option to send metrics about the pipeline to the backend (requires
-    setting the DD_API_KEY environment variable).
     """
+    project_name = "DataDog/datadog-agent"
 
     try:
-        messages_to_send, job_failure_stats = generate_failure_data()
+        failed_jobs = get_failed_jobs(project_name, os.getenv("CI_PIPELINE_ID"))
+        messages_to_send = generate_failure_messages(project_name, failed_jobs)
     except Exception as e:
         buffer = io.StringIO()
         print(base_message("datadog-agent", "is in an unknown state"), file=buffer)
@@ -472,14 +460,16 @@ def notify(_, notification_type="merge", print_to_stdout=False, send_stats=False
         traceback.print_exc()
         raise Exit(code=1)
 
-    # From the job failure stats, set whether the pipelin succeeded or failed and craft the
+    # From the job failures, set whether the pipeline succeeded or failed and craft the
     # base message that will be sent.
-    if job_failure_stats:  # At least one job failed
+    if failed_jobs:  # At least one job failed
         header_icon = ":host-red:"
         state = "failed"
+        coda = "If there is something wrong with the notification please contact #agent-platform"
     else:
         header_icon = ":host-green:"
         state = "succeeded"
+        coda = ""
 
     header = ""
     if notification_type == "merge":
@@ -494,61 +484,84 @@ def notify(_, notification_type="merge", print_to_stdout=False, send_stats=False
         message.base_message = base
         if owner not in GITHUB_SLACK_MAP.keys():
             message.base_message += UNKNOWN_OWNER_TEMPLATE.format(owner=owner)
-        message.coda = "If there is something wrong with the notification please contact #agent-platform"
+        message.coda = coda
         if print_to_stdout:
             print(f"Would send to {channel}:\n{str(message)}")
         else:
             send_slack_message(channel, str(message))  # TODO: use channel variable
 
-    # Send statistics, if enabled.
-    if send_stats:
-        if not os.environ.get("DD_API_KEY"):
-            print("DD_API_KEY environment variable not set, cannot send pipeline metrics to the backend")
-            raise Exit(code=1)
 
-        timestamp = int(datetime.now().timestamp())
-        series = []
+@task
+def send_stats(_, print_to_stdout=False):
+    """
+    Send statistics to Datadog for the current pipeline. CI-only task.
+    Use the --print-to-stdout option to test this locally, without sending
+    data points to Datadog.
+    """
+    project_name = "DataDog/datadog-agent"
 
-        # This stores the reason why a pipeline ultimately failed.
-        # The goal is to have a statistic of the number of pipelines that fail
-        # only due to infrastructure failures.
-        global_failure_reason = None
-        for failure_type, failure_reasons in job_failure_stats.items():
-            if failure_type == FailedJobType.JOB_FAILURE:
-                global_failure_reason = FailedJobType.JOB_FAILURE.name
-            elif failure_type == FailedJobType.INFRA_FAILURE and not global_failure_reason:
-                global_failure_reason = FailedJobType.INFRA_FAILURE.name
+    try:
+        job_failure_stats = get_failed_jobs_stats(project_name, os.getenv("CI_PIPELINE_ID"))
+    except Exception as e:
+        print("Found exception when generating statistics:")
+        traceback.print_exc(limit=-1)
+        raise Exit(code=1)
 
-            for failure_reason, count in failure_reasons.items():
-                # This allows getting stats on the number of jobs that fail due to infrastructure
-                # issues vs. other failures, and have a per-pipeline ratio of infrastructure failures.
-                series.append(
-                    create_count(
-                        metric_name="datadog.ci.job_failures",
-                        timestamp=timestamp,
-                        value=count,
-                        tags=["repository:datadog-agent", f"type:{failure_type.name}", f"reason:{failure_reason.name}"],
-                    )
+    if not os.environ.get("DD_API_KEY"):
+        print("DD_API_KEY environment variable not set, cannot send pipeline metrics to the backend")
+        raise Exit(code=1)
+
+    timestamp = int(datetime.now().timestamp())
+    series = []
+
+    # This stores the reason why a pipeline ultimately failed.
+    # The goal is to have a statistic of the number of pipelines that fail
+    # only due to infrastructure failures.
+    global_failure_reason = None
+    for failure_type, failure_reasons in job_failure_stats.items():
+        if failure_type == FailedJobType.JOB_FAILURE:
+            global_failure_reason = FailedJobType.JOB_FAILURE.name
+        elif failure_type == FailedJobType.INFRA_FAILURE and not global_failure_reason:
+            global_failure_reason = FailedJobType.INFRA_FAILURE.name
+
+        for failure_reason, count in failure_reasons.items():
+            # This allows getting stats on the number of jobs that fail due to infrastructure
+            # issues vs. other failures, and have a per-pipeline ratio of infrastructure failures.
+            series.append(
+                create_count(
+                    metric_name="datadog.ci.job_failures",
+                    timestamp=timestamp,
+                    value=count,
+                    tags=["repository:datadog-agent", f"type:{failure_type.name}", f"reason:{failure_reason.name}"],
                 )
-
-        pipeline_tags = ["repository:datadog-agent", f"status:{state}"]
-        if global_failure_reason:  # Only set the reason if the pipeline fails
-            pipeline_tags.append(f"reason:{global_failure_reason}")
-
-        series.append(
-            create_count(
-                metric_name="datadog.ci.pipelines",
-                timestamp=timestamp,
-                value=1,
-                tags=pipeline_tags,
             )
-        )
 
+    if job_failure_stats:  # At least one job failed
+        pipeline_state = "failed"
+    else:
+        pipeline_state = "succeeded"
+
+    pipeline_tags = ["repository:datadog-agent", f"status:{pipeline_state}"]
+    if global_failure_reason:  # Only set the reason if the pipeline fails
+        pipeline_tags.append(f"reason:{global_failure_reason}")
+
+    series.append(
+        create_count(
+            metric_name="datadog.ci.pipelines",
+            timestamp=timestamp,
+            value=1,
+            tags=pipeline_tags,
+        )
+    )
+
+    if not print_to_stdout:
         response = send_metrics(series)
         if response["errors"]:
             print(f"Error(s) while sending pipeline metrics to the Datadog backend: {response['errors']}")
             raise Exit(code=1)
         print(f"Sent pipeline metrics: {series}")
+    else:
+        print(f"Would send: {series}")
 
 
 def _init_pipeline_schedule_task():
